a good problem about binary search
we didnt solve it during contest. After studying others method for finding peak in mountain array we solve this problem ....
the difficult part for me is how to find out peak in mountain array, the
solution is just using binary search to find out the biggest number in
the array. we should practice it again and study binary search further