[Locked Problem]
after refer others method pass the test , the concept is Traverse the matrix. For each building, use BFS to compute the shortest distance from each '0' to
this building. After we do this for all the buildings, we can get the sum of shortest distance
from every '0' to all reachable buildings. we should practice it again
before go to interview