we didnt come out with critical concept which is
Assuming restriction for 1 is 2 times and we meet the case axxb (a, b are indexes of uncertain values). We all know we can not take 1 at index-b when xx = '11', these are all invaild so we need to remove them.

First of all, the total possible cases of dp[b][1] are sum of dp[b-1][1~6]
for index-b if we want to choose 1 so we need to consider the case that the two consecutive number before b is '11'. Also we need to be careful about value at the index-a. It shouldn't be '1'. In short we need to remove all possible cases of a11 and a is not 1.
The transition equation = dp[b][1] = sum(dp[b-1][1~6]) - sum(dp[a][2~6(except 1)])
after hint, beat 98%