greedy problem but we did not solve it in test
we need to practice it again ...

How can we know that this is the min number of steps?

First, We are not swapping untill we are getting chars at their original position.

Second, Since we are bringing character to its original position (if it was not),
that was the minimum effort (steps) we will have to put in since there is no other way of bringing it to original position with only adjacent swaps.

And now we can apply same thing to the rest of the string.
Plus we are swapping end to start, which shifts characters towards the end, and ultimately close to their original position,
because everything before that is already in their original position.