beat 30%
the fatest solution uses rank() function to change id and order it again
to generate answer ... very clever solution
The RANK() function assigns a rank to each row within the partition of a result set. The rank of a row is specified by one plus the number of ranks that come before it.
we should learn rank function and practice it again